fix: osgi packages - move api packages into separate package to make sure there are not import conflicts between core and api
