+++ aem-react-api/src/main/java/com/sinnerschrader/aem/reactapi/typescript/ExportTs.java
@@ -1,4 +1,4 @@
-package com.sinnerschrader.aem.react.typescript;
+package com.sinnerschrader.aem.reactapi.typescript;
 
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
