+++ core/src/main/java/com/sinnerschrader/aem/react/json/StringSerializer.java
@@ -14,6 +14,7 @@ import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.ser.ContextualSerializer;
+import com.sinnerschrader.aem.reactapi.json.NoResourceMapping;
 
 public class StringSerializer extends JsonSerializer<String> implements ContextualSerializer {
