+++ core/src/test/java/com/sinnerschrader/aem/react/json/ObjectMapperFactoryTest.java
@@ -17,6 +17,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.sinnerschrader.aem.reactapi.json.NoResourceMapping;
 
 @RunWith(MockitoJUnitRunner.class)
 public class ObjectMapperFactoryTest {
