+++ generate-typescript-maven-plugin/src/main/test/com/sinnerschrader/aem/react/tsgenerator/generator/BaseModel.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonSubTypes.Type;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
-import com.sinnerschrader.aem.react.typescript.ExportTs;
+import com.sinnerschrader.aem.reactapi.typescript.ExportTs;
 
 import lombok.Data;
