+++ generate-typescript-maven-plugin/src/main/test/com/sinnerschrader/aem/react/tsgenerator/generator/Sub2.java
@@ -1,6 +1,6 @@
 package com.sinnerschrader.aem.react.tsgenerator.generator;
 
-import com.sinnerschrader.aem.react.typescript.ExportTs;
+import com.sinnerschrader.aem.reactapi.typescript.ExportTs;
 
 import lombok.Data;
