+++ generate-typescript-maven-plugin/src/main/test/com/sinnerschrader/aem/react/tsgenerator/generator/TestArrayModel.java
@@ -1,7 +1,7 @@
 package com.sinnerschrader.aem.react.tsgenerator.generator;
 
-import com.sinnerschrader.aem.react.typescript.Element;
-import com.sinnerschrader.aem.react.typescript.ExportTs;
+import com.sinnerschrader.aem.reactapi.typescript.Element;
+import com.sinnerschrader.aem.reactapi.typescript.ExportTs;
 
 import lombok.Getter;
