+++ generate-typescript-maven-plugin/src/main/test/com/sinnerschrader/aem/react/tsgenerator/generator/TestGetterListModel.java
@@ -2,8 +2,8 @@ package com.sinnerschrader.aem.react.tsgenerator.generator;
 
 import java.util.List;
 
-import com.sinnerschrader.aem.react.typescript.Element;
-import com.sinnerschrader.aem.react.typescript.ExportTs;
+import com.sinnerschrader.aem.reactapi.typescript.Element;
+import com.sinnerschrader.aem.reactapi.typescript.ExportTs;
 
 import lombok.Getter;
