+++ generate-typescript-maven-plugin/src/main/test/com/sinnerschrader/aem/react/tsgenerator/generator/TestMapModel.java
@@ -2,8 +2,8 @@ package com.sinnerschrader.aem.react.tsgenerator.generator;
 
 import java.util.Map;
 
-import com.sinnerschrader.aem.react.typescript.Element;
-import com.sinnerschrader.aem.react.typescript.ExportTs;
+import com.sinnerschrader.aem.reactapi.typescript.Element;
+import com.sinnerschrader.aem.reactapi.typescript.ExportTs;
 
 import lombok.Getter;
